Removed dependancy on <eku\io\_io.h>
Added facility to test only anions and only cations instead of checking complete salt.
This had to be done because distribution of ions was not uniform.
For eg. there are many nitrates in ANAL_SALTS because they are always soluble while sulphides were uncommon as they are usually insoluble.

Added ability to convert 'ion(s)' to 'ion(aq)' in dissociating medium in function tube_t::react_prep()